clock setting improvements

+++ my_drivers/Inc/my_tims.h
@@ -8,7 +8,7 @@
 #ifndef MY_TIMS_H_
 #define MY_TIMS_H_
 
-void config_my_tim16	( uint16_t ) ;
+void config_my_tim16	( uint32_t ) ;
 void start_my_tim16 	( uint16_t ) ;
 void stop_my_tim16 		( void ) ;
 void off_my_tim16 		( void ) ;
